Order is important in require/define
If it's the first argument in your module function, it needs to be the
first item in your array of dependencies. In this case, we access
Angular through the global variable and not a module, so we don't need
to get to it through an argument, and it can go last.